% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.r
\name{assortativity}
\alias{assortativity}
\title{Calculating network assortativity}
\usage{
assortativity(network)
}
\arguments{
\item{network}{matrix indicating network structure (symmetric)}
}
\value{
Numeric value
}
\description{
The function calculates the assortativity of a network, 
that measures the preference of interactions between similar nodes. 
As in most literature, 'similarity' is here defined in terms of node degrees.
}
\examples{
network <- matrix(c(1,0,1,0,1,1,0,1,1),nrow=3,byrow=T)
genes <- c('gene1','gene2','gene3')
rownames(network) <- genes
colnames(network) <- genes
assort_value <- assortativity(network)

}
\keyword{assortativity}
\keyword{network}
\keyword{properties}
\keyword{topology}

