% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.mouse.R
\docType{data}
\name{attr.mouse}
\alias{attr.mouse}
\title{Mouse GENCODE annotations (M7)}
\format{
A data frame with 46517 rows and 10 variables:
\describe{
  \item{chr}{chromosome}
  \item{start}{chromosomal start position, in base pairs}
  \item{end}{chromosomal end position, in base pairs}
  \item{strand}{chromosomal strand, + or - }
  \item{un}{unknown}
  \item{ensemblID}{ENSEMBL identifier}
  \item{type}{type of transcript}
  \item{stat}{status of transcript}
  \item{name}{HUGO identifier}
  \item{entrezID}{Entrez identifier}
 
}
 @source \url{ftp://ftp.sanger.ac.uk/pub/gencode/Gencode_mouse/release_M7/}
}
\description{
A dataset containing identifiers for gene transcripts
}
