% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_multifunc.R
\name{auc_multifunc}
\alias{auc_multifunc}
\title{Calculating AUC for functional groups from ranked lists}
\usage{
auc_multifunc(annotations)
}
\arguments{
\item{annotations}{binary matrix indicating which list elements are in which functional groups.}
}
\value{
aucs array of aucs for each group in annotations
}
\description{
The function calculates the AUC for a functional group 
analytically using an optimal ranked list of genes that 
indicates association between genes and groups.
}
\examples{
annotations <- matrix( sample(c(0,1), 100, replace=TRUE),ncol=10,nrow=10)
rownames(annotations) = paste('gene', 1:10, sep='')
colnames(annotations) = paste('label', 1:10, sep='')
aurocs_mf <- auc_multifunc(annotations)

}
\keyword{AUC}
\keyword{evalutation}
\keyword{multifunctionality}

