% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_singlelist.R
\name{auc_singlelist}
\alias{auc_singlelist}
\title{Area under the receiver operating characteristic curve}
\usage{
auc_singlelist(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
auroc Numeric value
}
\description{
The function calculates the area under the receiver operating characteristic (ROC)
curve analytically
}
\examples{
labels <- c(rep(0,10))
labels[c(1,3,5)] <- 1 
scores <- 10:1
auroc <- auc_singlelist(scores, labels)

}
\keyword{ROC}
\keyword{analytic}
\keyword{area}
\keyword{characteristic}
\keyword{metric}
\keyword{operating}
\keyword{receiver}

