% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auprc.r
\name{auprc}
\alias{auprc}
\title{Area under the precision recall curve}
\usage{
auprc(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
auprc Numeric value
}
\description{
The function calculates the area under the precision-recall curve
}
\keyword{area}
\keyword{metric}
\keyword{precision-recall}

