% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auroc_analytic.r
\name{auroc_analytic}
\alias{auroc_analytic}
\title{Area under the receiver operating characteristic curve}
\usage{
auroc_analytic(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
auroc Numeric value
}
\description{
The function calculates the area under the receiver operating characteristic (ROC)
curve analytically
}
\keyword{ROC}
\keyword{analytic}
\keyword{area}
\keyword{characteristic}
\keyword{metric}
\keyword{operating}
\keyword{receiver}

