% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogrid.R
\docType{data}
\name{biogrid}
\alias{biogrid}
\title{BIOGRID v3.4.126}
\format{A data frame with 211506 rows and 2 variables:
\describe{
  \item{entrezID_A}{List of Entrez identifiers, interactor A}
  \item{entrezID_B}{List of Entrez identifiers, interactor B}
 
}
 @source \url{http://thebiogrid.org/}}
\description{
A data frame containing protein-protein interactions
}

