% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_gene_ids.r
\name{convert_gene_ids}
\alias{convert_gene_ids}
\title{Gene identifier conversion tool}
\usage{
convert_gene_ids(genelist, to = "Entrez", from = "Ensembl",
  species = "9606")
}
\arguments{
\item{genelist}{array of gene IDs}

\item{to}{string identifier, default Entrez}

\item{from}{string identifier, default Ensembl}

\item{species}{string identifier, default 9606 (human)}
}
\value{
genelist_converted array of gene IDs
}
\description{
The function maps between gene ids
}
\keyword{Ensembl}
\keyword{Entrez}
\keyword{HUGO}
\keyword{IDs}
\keyword{Symbol}
\keyword{gene}

