% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_neighbor_voting.R
\docType{data}
\name{example_neighbor_voting}
\alias{example_neighbor_voting}
\title{Example of binary network}
\format{\describe{
  \item{entrezID}{chromosomal start position, in base pairs}
  \item{name}{HUGO gene identifier}
  \item{species}{species}
  \item{disease}{disease}
 
}}
\description{
This dataset includes
}

