% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_network.R
\name{filter_network}
\alias{filter_network}
\title{Filter on matrix}
\usage{
filter_network(network, flag = 1, min = 0, max = 1, ids = NA)
}
\arguments{
\item{network}{numeric matrix}

\item{flag}{numeric 1 for row filtering, 2 for column filtering}

\item{min}{numeric value}

\item{max}{numeric value}

\item{ids}{array to filter on}
}
\value{
network numeric matrix
}
\description{
The function filters out the rows or columns of a matrix such 
that the size of the group is exclusively between given min 
and max values
}
\examples{
net <- matrix( rnorm(10000), nrow=100)
filt_net <- filter_network(net,1,10,100)

}
\keyword{filter}
\keyword{network}
\keyword{rows}

