% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_network_cols.R
\name{filter_network_cols}
\alias{filter_network_cols}
\title{Filter on columns}
\usage{
filter_network_cols(network, min, max)
}
\arguments{
\item{network}{numeric matrix}

\item{min}{numeric value}

\item{max}{numeric value}
}
\value{
network numeric matrix
}
\description{
The function filters out the columns of a matrix such that the size of the group
is exclusively between given min and max values
}
\keyword{filter}
\keyword{network}
\keyword{rows}

