% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmeasure.r
\name{fmeasure}
\alias{fmeasure}
\title{Fmeasure of precision-recall}
\usage{
fmeasure(recall, precis, beta = 1)
}
\arguments{
\item{recall}{numeric array}

\item{precis}{numeric array}

\item{beta}{numeric value, default is 1}
}
\value{
fmeasure Numeric value
}
\description{
The function calculates fmeasure for a given beta of a precision-recall curve
}
\keyword{fmeasure}
\keyword{precision-recall}

