% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes.R
\name{genes}
\alias{genes}
\title{Genes from BIOGRID v3.4.126}
\format{Array
\describe{
  \item{genes}{List of Entrez identifiers}
 
}
 @source \url{http://thebiogrid.org/}}
\description{
An array containing identifiers for genes in biogrid
}

