% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auc.r
\name{get_auc}
\alias{get_auc}
\title{Calculates the area under a curve}
\usage{
get_auc(x, y)
}
\arguments{
\item{x}{numeric array}

\item{y}{numeric array}
}
\value{
auc numeric value
}
\description{
The function calculates the area under the curve defined by x and y
}
\keyword{a}
\keyword{area}
\keyword{curve}
\keyword{mean}
\keyword{rolling}
\keyword{under}

