% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biogrid.R
\name{get_biogrid}
\alias{get_biogrid}
\title{Downloading and filtering BIOGRID}
\usage{
get_biogrid(species = "9606", version = "3.4.131",
  interactions = "physical")
}
\arguments{
\item{species}{numeric taxon of species}

\item{version}{string of biogrid version}

\item{interactions}{string stating either physical or genetic interactions}
}
\value{
biogrid data.frame with interactions
}
\description{
The function downloads the specifed version of biogrid for a particular taxon
}
\keyword{biogrid}
\keyword{download}

