% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.r
\name{get_counts}
\alias{get_counts}
\title{Get counts}
\usage{
get_counts(hist)
}
\arguments{
\item{hist}{histogram}
}
\value{
x,y
}
\description{
The function formats the count distribution from the histogram function
}
\keyword{histogram}
\keyword{plot}

