% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression_data_gemma.r
\name{get_expression_data_gemma}
\alias{get_expression_data_gemma}
\title{Obtain expression matrix from the GEMMA database}
\usage{
get_expression_data_gemma(gseid, filtered = "true")
}
\arguments{
\item{gseid}{GEO ID of the expression experiment}

\item{filtered}{flag to indicate whether or not the data is QC}
}
\value{
list of genes and the expression matrix
}
\description{
The function downloads and parses the expression matrix from the GEMMA 
database, specified by the GEO ID
}
\keyword{GEMMA}
\keyword{GEO}
\keyword{GSE}
\keyword{experiment}
\keyword{expression}

