% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expression_matrix_from_GEO.R
\name{get_expression_matrix_from_GEO}
\alias{get_expression_matrix_from_GEO}
\title{Obtain expression matrix from GEO database}
\usage{
get_expression_matrix_from_GEO(gseid)
}
\arguments{
\item{gseid}{GEO ID of the expression experiment}
}
\value{
list of genes and the expression matrix
}
\description{
The function downloads and parses the expression matrix from the GEO file 
specified by the GEO ID
}
\keyword{GEO}
\keyword{GSE}
\keyword{experiment}
\keyword{expression}

