% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenocarta.R
\name{get_phenocarta}
\alias{get_phenocarta}
\title{Downloading and filtering Phenocarta}
\usage{
get_phenocarta(species = "human", type = "all")
}
\arguments{
\item{species}{string}

\item{type}{string}
}
\value{
data data.frame with phenocarta data
}
\description{
The function downloads the latest version of phenocarta
}
\keyword{download}
\keyword{phenocarta}
