% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.human.R
\docType{data}
\name{GO.human}
\alias{GO.human}
\title{GO - human}
\format{A data frame with 2511938 rows and 4 variables:
\describe{
  \item{name}{chromosome}
  \item{entrezID}{chromosomal start position, in base pairs}
  \item{GO}{chromosomal end position, in base pairs}
  \item{evidence}{chromosomal strand, + or - }
 
}
 @source \url{http://geneontology.org/}}
\description{
A dataset of the gene GO associations
}

