% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.voc.R
\docType{data}
\name{GO.voc}
\alias{GO.voc}
\title{Gene ontology vocabulary}
\format{
A data frame with 42266 rows and 3 variables:
\describe{
  \item{GOID}{GO identifier}
  \item{term}{GO description}
  \item{domain}{GO domain}
 
}
 @source \url{http://geneontology.org/}
}
\description{
A dataset of the gene ontology vocabulary
}
