% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_annotations.R
\name{make_annotations}
\alias{make_annotations}
\title{Creating gene annotations}
\usage{
make_annotations(data, listA, listB)
}
\arguments{
\item{data}{2-column matrix, each row a pair indicating a relationship or interaction}

\item{listA}{string array of genes}

\item{listB}{string array of labels/functions}
}
\value{
net matrix binary
}
\description{
The function annotates a list of genes according to a given ontology. 
It creates a binary matrix associating genes (rows) with labels (columns).
}
\examples{
data <- cbind(genes=c('gene1','gene1','gene2','gene3','gene3','gene3'),
GO=c('GO:1','GO:3','GO:2','GO:1','GO:2','GO:3'))
listA <- c('gene1','gene2','gene3')
listB <- c('GO:1','GO:2','GO:3')
annotations <- make_annotations(data,listA,listB)

 
}
\keyword{annotations}
\keyword{function}
\keyword{gene}
\keyword{labels}
\keyword{ontology}
\keyword{to}

