% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transparent.R
\name{make_transparent}
\alias{make_transparent}
\title{Make a color transparent (Taken from an answer on StackOverflow by Nick Sabbe)}
\usage{
make_transparent(someColor, alpha = 100)
}
\arguments{
\item{someColor}{color number, string or hexidecimal code}

\item{alpha}{numeric transparency}
}
\value{
someColor rgb
}
\description{
Make a color transparent (Taken from an answer on StackOverflow by Nick Sabbe)
}

