% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_voting.R
\name{neighbor_voting}
\alias{neighbor_voting}
\title{Evaluating Gene Function Prediction}
\usage{
neighbor_voting(genes.labels, network, nFold = 3, output = "AUC")
}
\arguments{
\item{genes.labels}{numeric array}

\item{network}{numeric array symmetric, gene-by-gene matrix}

\item{nFold}{numeric value, default is 3}

\item{output}{string, default is AUC}
}
\value{
scores numeric matrix
}
\description{
The function performs gene function prediction based on 'guilt by association' 
using cross validation ([1]). Performance and significance are evaluated by 
calculating the AUROC or AUPRC of each functional group.
}
\examples{

data <- cbind(geneA=c('gene1','gene2'),geneB=c('gene3','gene3'))
genelist <- c('gene1','gene2','gene3')
network <- make_gene_network(data,genelist)
ontology <- cbind(genes=c('gene1','gene1','gene2','gene3','gene3','gene3'), 
GO=c('GO:1','GO:3','GO:2','GO:1','GO:2','GO:3'))
goterms <- c('GO:1','GO:2','GO:3')
nFold <- 3 
annotations <- make_annotations(ontology,genelist,goterms)
network <- make_gene_network(data,genelist)
roc.scores <- neighbor_voting(annotations,network,nFold)

}
\keyword{association}
\keyword{by}
\keyword{cross}
\keyword{evaluation}
\keyword{function}
\keyword{gene}
\keyword{guilt}
\keyword{neighbor}
\keyword{prediction}
\keyword{validation}
\keyword{voting}

