% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_degree.r
\name{node_degree}
\alias{node_degree}
\title{Calculate node degree}
\usage{
node_degree(net)
}
\arguments{
\item{net}{numeric matrix}
}
\value{
node_degree numeric array
}
\description{
The function calculates the node degree of a network
}
\keyword{degree}
\keyword{metric}
\keyword{network}
\keyword{node}
\keyword{topology}

