% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho.R
\docType{data}
\name{ortho}
\alias{ortho}
\title{Gene orthologs}
\format{List orthologs for 5 species
\describe{
  \item{dros}{List of Entrez identifiers, Drosophila}
  \item{celeg}{List of Entrez identifiers, C. elegans}
  \item{yeast}{List of Entrez identifiers, Yeast}
  \item{mouse}{List of Entrez identifiers, Mouse}
  \item{zf}{List of Entrez identifiers, Zebrafish}
 
}
 @source \url{http://useast.ensembl.org/index.html/}}
\description{
A list containing identifiers for the subsets of gene orthologs
}

