% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Phenocarta}
\format{A data frame with 142272 rows and 4 variables:
\describe{
  \item{entrezID}{chromosomal start position, in base pairs}
  \item{name}{HUGO gene identifier}
  \item{species}{species}
  \item{disease}{disease}
 
}
 @source \url{http://www.chibi.ubc.ca/Gemma/phenotypes.html}}
\description{
A dataset of gene disease associations
}
