% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_densities.r
\name{plot_densities}
\alias{plot_densities}
\title{Plot densities}
\usage{
plot_densities(hists, id, col = "lightgrey", xlab = "log2 expression FPKM",
  ylab = "Density", mode = "hist", ...)
}
\arguments{
\item{aucA}{numeric array of aucs}

\item{aucB}{numeric array of aucs}
}
\description{
The function plots multiple density curves and compares their modes
}
\keyword{plot}

