% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network_heatmap.R
\name{plot_network_heatmap}
\alias{plot_network_heatmap}
\title{Plot network heatmap}
\usage{
plot_network_heatmap(net, colrs)
}
\arguments{
\item{net}{a numeric matrix of edge weights}

\item{colrs}{a range of colors to plot the network}
}
\value{
null
}
\description{
The function draws a heatmap to visualize a network
}
\examples{
network <- cor(matrix( rnorm(10000), nrow=100))
plot_network_heatmap(network)

}
\keyword{heatmap}
\keyword{image}
\keyword{network}
\keyword{plot}

