% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prc.R
\name{plot_prc}
\alias{plot_prc}
\title{Plot precision recall curve}
\usage{
plot_prc(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
prc numeric arrays
}
\description{
The function calculates the precision and recall and plots the curve
}
\examples{
labels <- c(rep(0,10))
labels[c(1,3,5)] <- 1 
scores <- 10:1
roc <- plot_prc(scores, labels)


}
\keyword{PRC}
\keyword{metric}
\keyword{plot}
\keyword{precision}
\keyword{recall}

