% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot receiver operating characteristic curve}
\usage{
plot_roc(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
FPR,TPR numeric arrays
}
\description{
The function calculates the FPR and TRPR for the receiver
operating characteristic (ROC) and plots the curve
}
\examples{
labels <- c(rep(0,10))
labels[c(1,3,5)] <- 1 
scores <- 10:1
roc <- plot_roc(scores, labels)


}
\keyword{FPR}
\keyword{ROC}
\keyword{TPR}
\keyword{characteristic}
\keyword{metric}
\keyword{operating}
\keyword{receiver}

