% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc_overlay.r
\name{plot_roc_overlay}
\alias{plot_roc_overlay}
\title{Plot ROC overlay}
\usage{
plot_roc_overlay(scores.mat, labels.mat, nbins = 100)
}
\arguments{
\item{scores.mat}{numeric array}

\item{labels.mat}{numeric array}

\item{nbins}{numeric value}
}
\value{
list of Z(matrix) and roc_sum (average ROC curve)
}
\description{
The function plots a density overlay of ROCs given the scores and labels
}
\keyword{ROC}
\keyword{overlay}
\keyword{plot}

