% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_value_compare.R
\name{plot_value_compare}
\alias{plot_value_compare}
\title{Plot value comparisons}
\usage{
plot_value_compare(aucA, aucB, xlab = "AUROC", ylab = "AUROC", xlim = c(0,
  1), ylim = c(0, 1))
}
\arguments{
\item{aucA}{numeric array of aucs}

\item{aucB}{numeric array of aucs}

\item{xlab}{string label}

\item{ylab}{string label}

\item{xlim}{range of values for xaxis}

\item{ylim}{range of values for yaxis}
}
\value{
null
}
\description{
The function plots a scatter
}
\keyword{plot}

