% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predictions}
\alias{predictions}
\title{Performing Gene Function Prediction}
\usage{
predictions(genes.labels, network)
}
\arguments{
\item{genes.labels}{numeric array}

\item{network}{numeric array symmetric, gene-by-gene matrix}
}
\value{
scores numeric matrix
}
\description{
The function performs gene function prediction on the whole data set
using the 'guilt by association'-principle ([1]).
}
\examples{
data <- cbind(geneA=c('gene1','gene2'),geneB=c('gene3','gene3'))
genelist <- c('gene1','gene2','gene3')
network <- make_gene_network(data,genelist)
ontology <- cbind(genes=c('gene1','gene1','gene2','gene3','gene3','gene3'), 
GO=c('GO:1','GO:3','GO:2','GO:1','GO:2','GO:3'))
goterms <- c('GO:1','GO:2','GO:3')
nFold <- 3 
annotations <- make_annotations(ontology,genelist,goterms)
network <- make_gene_network(data,genelist)
predicted_functions <- predictions(network,annotations)

}
\keyword{association}
\keyword{by}
\keyword{function}
\keyword{gene}
\keyword{guilt}
\keyword{neighbor}
\keyword{prediction}
\keyword{voting}

