% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repmat.r
\name{repmat}
\alias{repmat}
\title{Rep function for matrices}
\usage{
repmat(X, m, n)
}
\arguments{
\item{X}{numeric matrix}

\item{m}{numeric value, repeat rows m times}

\item{n}{numeric value, repeat columns n times}
}
\value{
list of genes and the expression matrix
}
\description{
The function generates a matrix by binding the columns and rows
}
\examples{

}
\keyword{matrix}
\keyword{repeat}
\keyword{repmat}

