% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auprc.R
\name{auprc}
\alias{auprc}
\title{Area under the precision recall curve}
\usage{
auprc(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
auprc Numeric value
}
\description{
The function calculates the area under the precision-recall curve
}
\examples{
labels <- c(rep(0,10))
labels[c(1,3,5)] <- 1 
scores <- 10:1
auprc <- auprc(scores, labels)


}
\keyword{area}
\keyword{metric}
\keyword{precision-recall}
