% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_semantic_similarity_network.R
\name{build_semantic_similarity_network}
\alias{build_semantic_similarity_network}
\title{Builds a semantic similarity network}
\usage{
build_semantic_similarity_network(genes.labels, genes)
}
\arguments{
\item{genes.labels}{matrix with rows as genes and columns as a function/label}

\item{genes}{array of gene IDs}
}
\value{
net Numeric value
}
\description{
The function builds a semantic similarity network given a data and labels
}
\examples{
genes.labels <- matrix( sample(c(0,1), 100, replace=TRUE),ncol=10,nrow=10)
rownames(genes.labels) <- 1:10
genes <- 1:10
net <- build_semantic_similarity_network(genes.labels, genes)

}
\keyword{jaccard}
\keyword{network}
\keyword{semantic}
\keyword{similarity}

