% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multifunc.R
\name{calculate_multifunc}
\alias{calculate_multifunc}
\title{Performing multifunctionality analysis}
\usage{
calculate_multifunc(genes.labels)
}
\arguments{
\item{genes.labels}{Annotation matrix}
}
\value{
Returns matrix with evaluation of gene function prediction by given labels:
}
\description{
The function performs multifunctionality analysis ([1]) 
for a set of annotated genes and creates a rank based 
optimallist. For annotations use an ontology that is 
large enough to serve as a prior (e.g. GO, Phenocarta).
}
\examples{
data <- cbind(genes=c('gene1','gene1','gene2','gene3','gene3','gene3'), 
GO=c('GO:1','GO:3','GO:2','GO:1','GO:2','GO:3'))
listA <- c('gene1','gene2','gene3')
listB <- c('GO:1','GO:2','GO:3')
annotations <- make_annotations(data,listA,listB)
multifunc_assessment <- calculate_multifunc(annotations)

}
\author{

}
\keyword{evaluation}
\keyword{function}
\keyword{gene}
\keyword{ontology}
\keyword{prediction}

