% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_smoother.R
\name{conv_smoother}
\alias{conv_smoother}
\title{Plot smoothed curve}
\usage{
conv_smoother(X, Y, window, xlab = "", ylab = "", raw = FALSE)
}
\arguments{
\item{X}{numeric array}

\item{Y}{numeric array}

\item{window}{numeric value indicating size of window to use}

\item{xlab}{string of x-axis label}

\item{ylab}{string of y-axis label}

\item{raw}{boolean}
}
\value{
smoothed X,Y and std Y matrix
}
\description{
The function plots a smoothed curve using the \code{\link{convolve}} function.
}
\examples{
x <- 1:1000
y <- rnorm(1000)
conv <- conv_smoother(x,y,10)

}
\keyword{plot}
\keyword{smooth}

