% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmeasure.R
\name{fmeasure}
\alias{fmeasure}
\title{Fmeasure of precision-recall}
\usage{
fmeasure(recall, precis, beta = 1)
}
\arguments{
\item{recall}{numeric array}

\item{precis}{numeric array}

\item{beta}{numeric value, default is 1}
}
\value{
fmeasure Numeric value
}
\description{
The function calculates fmeasure for a given beta of a precision-recall curve
}
\examples{
labels <- c(rep(0,10))
labels[c(1,3,5)] <- 1 
scores <- 10:1
prc <- get_prc(scores, labels)
fm <- fmeasure(prc[,1], prc[,2])

}
\keyword{fmeasure}
\keyword{precision-recall}

