% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auc.R
\name{get_auc}
\alias{get_auc}
\title{Calculates the area under a curve}
\usage{
get_auc(x, y)
}
\arguments{
\item{x}{numeric array}

\item{y}{numeric array}
}
\value{
auc numeric value
}
\description{
The function calculates the area under the curve defined by x and y
}
\examples{
x <- 1:100
y <- 1:100 
auc <- get_auc(x,y)
 
}
\keyword{a}
\keyword{area}
\keyword{curve}
\keyword{mean}
\keyword{rolling}
\keyword{under}
