% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{Get counts}
\usage{
get_counts(hist)
}
\arguments{
\item{hist}{histogram}
}
\value{
x,y
}
\description{
The function formats the count distribution from the histogram function
}
\examples{
x <- runif(1000)
counts <- get_counts( hist(x, plot=FALSE))

}
\keyword{histogram}

