% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_density.R
\name{get_density}
\alias{get_density}
\title{Get density}
\usage{
get_density(hist)
}
\arguments{
\item{hist}{histogram}
}
\value{
array
}
\description{
The function formats the density distribution from the histogram function
}
\examples{
x <- runif(1000)
density <- get_density( hist(x, plot=FALSE))

}
\keyword{histogram}
\keyword{plot}

