% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prc.r
\name{get_prc}
\alias{get_prc}
\title{Build precision-recall curve}
\usage{
get_prc(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
recall,precision numeric arrays
}
\description{
The function calculates the recall and precision
}
\keyword{precision}
\keyword{precsion-recall}
\keyword{recall}

