% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roc.R
\name{get_roc}
\alias{get_roc}
\title{Build receiver operating characteristic curve}
\usage{
get_roc(scores, labels)
}
\arguments{
\item{scores}{numeric array}

\item{labels}{binary array}
}
\value{
FPR,TPR numeric arrays
}
\description{
The function calculates the FPR and TRPR for the receiver 
operating characteristic (ROC)
}
\examples{
labels <- c(rep(0,10))
labels[c(1,3,5)] <- 1 
scores <- 10:1
roc <- get_roc(scores, labels)



}
\keyword{FPR}
\keyword{ROC}
\keyword{TPR}
\keyword{characteristic}
\keyword{metric}
\keyword{operating}
\keyword{receiver}

