% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.mouse.R
\docType{data}
\name{GO.mouse}
\alias{GO.mouse}
\title{GO - mouse}
\format{A data frame with 2086086 rows and 4 variables:
\describe{
  \item{name}{gene symbol}
  \item{entrezID}{entrez identifier}
  \item{GO}{gene ontology term ID}
  \item{evidence}{evidence code}
}
 @source \url{http://geneontology.org/}}
\description{
A dataset of the gene GO associations
}

