% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_genelist.R
\name{make_genelist}
\alias{make_genelist}
\title{Creating list of all genes in the data set.}
\usage{
make_genelist(gene_data_interacting)
}
\arguments{
\item{gene_data_interacting}{2-column matrix, each row a pair indicating a relationship or interaction}
}
\value{
list array of data labels
}
\description{
The function extracts the list of all genes in the data set
}
\examples{
gene.list <- paste('gene', 1:100, sep='')
data <- matrix(0,nrow=100, ncol=2)
data[,1] <- sample(gene.list, 50, replace=TRUE)
data[,2] <- sample(gene.list, 50, replace=TRUE)
genes <- make_genelist(data)

}
\keyword{extract}
\keyword{gene}
\keyword{list}
