% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density_compare.r
\name{plot_density_compare}
\alias{plot_density_compare}
\title{Plot density comparisons}
\usage{
plot_density_compare(aucA, aucB, col = "lightgrey",
  xlab = "AUC (neighbor voting)", ylab = "Density", mode = TRUE)
}
\arguments{
\item{aucA}{numeric array of aucs}

\item{aucB}{numeric array of aucs}

\item{col}{color of lines}

\item{xlab}{string label}

\item{ylab}{string label}

\item{mode}{boolean to plot mode or mean}
}
\description{
The function plots two density curves and compares their modes
}
\keyword{plot}

