% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GBA.R
\name{run_GBA}
\alias{run_GBA}
\title{Performing 'Guilt by Association' Analysis}
\usage{
run_GBA(network, labels, min = 20, max = 1000, nfold = 3)
}
\arguments{
\item{network}{numeric array symmetric, gene-by-gene matrix}

\item{labels}{numeric array}

\item{min}{numeric value to limit gene function size}

\item{max}{numeric value to limit gene function size}

\item{nfold}{numeric value, default is 3}
}
\value{
list roc.sub, genes, auroc
}
\description{
The function runs and evaluates gene function prediction based on the 
'guilt by association'-principle using neighbor voting (\code{\link{neighbor_voting}})
[1]. As a measure of performance and significance of results, AUCs of all 
evaluated functional groups are calculated.
}
\examples{

data <- cbind(geneA=c('gene1','gene2'),geneB=c('gene3','gene3'))
genelist <- c('gene1','gene2','gene3')
network <- make_gene_network(data,genelist)
ontology <- cbind(genes=c('gene1','gene1','gene2','gene3','gene3','gene3'), 
GO=c('GO:1','GO:3','GO:2','GO:1','GO:2','GO:3'))
goterms <- c('GO:1','GO:2','GO:3')
nFold <- 3 
annotations <- make_annotations(ontology,genelist,goterms)
network <- make_gene_network(data,genelist)
GBA_analysis <- run_GBA(network,labels)

}
\keyword{association}
\keyword{by}
\keyword{cross}
\keyword{evaluation}
\keyword{function}
\keyword{gene}
\keyword{guilt}
\keyword{neighbor}
\keyword{prediction}
\keyword{validation}
\keyword{voting}

